refactor how transformers are structured; remove @transformer decorator
