+++ robotidy/decorators.py
@@ -1,25 +1,6 @@
-import inspect
 import functools
 
 from robotidy.utils import node_within_lines
-
-
-TRANSFORMERS = set()
-
-
-def transformer(arg=None):
-    """Decorator for transformer class. Only decorated classes are loaded and used to transform the source code."""
-
-    # it allows to use transformer without ()
-    if inspect.isclass(arg):
-        TRANSFORMERS.add(arg.__name__)
-        return transformer()(arg)
-
-    def decorator(cls):
-        cls.is_transformer = True
-        return cls
-
-    return decorator
 
 
 def return_node_untouched(node):
